fixed for -Wconversion

+++ include/votca/xtp/atomcontainer.h
@@ -52,7 +52,7 @@ class AtomContainer {
 
   int getId() const { return _id; }
 
-  int size() const { return _atomlist.size(); }
+  long int size() const { return _atomlist.size(); }
 
   void push_back(const T& atom) {
     _atomlist.push_back(atom);
