+++ include/votca/xtp/couplingbase.h
@@ -72,7 +72,7 @@ inline void CouplingBase::CheckAtomCoordinates(
   const QMMolecule& atomsB = orbitalsB.QMAtoms();
   const QMMolecule& atomsAll = orbitalsAB.QMAtoms();
   bool coordinates_agree = true;
-  for (int i = 0; i < atomsAll.size(); i++) {
+  for (long int i = 0; i < atomsAll.size(); i++) {
     const QMAtom& dimer = atomsAll[i];
     const QMAtom* monomer = nullptr;
